@@ -28,12 +28,9 @@ allprojects {
 }
 
 ext {
-    compileSdkVersion = 25
-    buildToolsVersion = "25.0.0"
+    compileSdkVersion = 28
     minSdkVersion = 16
-    targetSdkVersion = 25
+    targetSdkVersion = 28
 
-    supportVersion = "25.1.0"
-
-    aaVersion = "3.3.2"
+    supportVersion = '1.0.0-beta01'
 }
