@@ -1,8 +1,8 @@
 package org.xdty.gallery.view;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 import org.xdty.gallery.fragment.ImageFragment;
 import org.xdty.gallery.model.Media;
